# Conflicts:
#	.idea/caches/build_file_checksums.ser
#	.idea/misc.xml
#	app/src/main/java/com/afaa/tanktrouble/battle/Position.java
#	app/src/main/java/com/afaa/tanktrouble/cannonball/Cannonball.java
#	app/src/main/java/com/afaa/tanktrouble/datahouse/BluetoothService.java
#	app/src/main/java/com/afaa/tanktrouble/datahouse/DataProtocol.java
#	app/src/main/java/com/afaa/tanktrouble/datahouse/DeviceChooserActivity.java
#	app/src/main/java/com/afaa/tanktrouble/datahouse/GameData.java
#	app/src/test/java/com/afaa/tanktrouble/ExampleUnitTest.java
#	build.gradle
#	gradle/wrapper/gradle-wrapper.properties

+++ app/src/main/java/com/afaa/tanktrouble/cannonball/Cannonball.java
@@ -241,7 +241,7 @@ public class Cannonball {
     }
 
     public Position getPosition() {
-        return new Position(mX, mY, mDeg);
+        return new Position((int) mX, (int) mY, mDeg);
     }
 
     public int getRadius() {
